+++ src/modules/meal/meal.utils.ts
@@ -1,6 +1,8 @@
 import { MealType } from '../../common/enums';
 import { ApiName } from '../redis/redis.types';
 import { IngredientType } from '../ingredient/ingredient.types';
+import { MealDocument } from '../../mongodb/documents/meal.document';
+import { DetailedMeal } from './meal.types';
 
 export function getQueryWithIngredientsAndMealType(ingredients: IngredientType[], type: MealType, apiName?: ApiName, apiKey?: string): string {
     const ingredientList = ingredients.sort().join(',');
@@ -19,4 +21,15 @@ export function getQueryWithIngredientsAndMealType(ingredients: IngredientType[]
     default:
         throw new Error('Unknown API name case.');
     }
+}
+
+export function proceedMealDocumentToMealDetails(meal: MealDocument): DetailedMeal {
+    const { id, imageUrl, ingredients, title } = meal;
+
+    return {
+        id,
+        imgUrl: imageUrl,
+        ingredients,
+        title
+    };
 }