+++ src/modules/redis/redis.types.ts
@@ -1,5 +1,6 @@
 export type TokenKey = `user:${string}:${'accessToken' | 'refreshToken'}`;
 
 export type MealResultQueryKey = `query:${ApiName}:${string}`;
+export type MealDetailsQueryKey = `meal-details:${string}`;
 
 export type ApiName = 'merged' | 'localmongo' | 'spoonacular';