+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.25.0] - 2024-02-15
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Filtering provided ingredients - only supported ingredients.
+
 ### Changed
 - [Marek Kurańda](https://github.com/mjkuranda): Ingredient as a string instead of object.